Merge

+++ Listrr/Services/TraktService.cs
@@ -1,10 +1,10 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-
-using Listrr.Data.Trakt;
+﻿using Listrr.Data.Trakt;
 using Listrr.Repositories;
 
 using Microsoft.AspNetCore.Identity;
+
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 using TraktNet.Objects.Get.Movies;
 using TraktNet.Objects.Get.Shows;
@@ -41,10 +41,12 @@ namespace Listrr.Services
             return model;
         }
 
-        public async Task Delete(TraktList model)
+        public async Task Delete(TraktList model, bool onlyLocal = true)
         {
             await _traktListDbRepository.Delete(model);
-            await _traktListApiRepository.Delete(model);
+
+            if(!onlyLocal)
+                await _traktListApiRepository.Delete(model);
         }
 
         public async Task<TraktList> Get(uint id, bool forceAPI = false)
